Intinialize psse converter
