+++ src/power_grid_model_io/data_types/__init__.py
@@ -6,6 +6,16 @@ Common data types used in the Power Grid Model project
 """
 
 from power_grid_model_io.data_types._data_types import ExtraInfo, ExtraInfoLookup, StructuredData
-from power_grid_model_io.data_types.tabular_data import LazyDataFrame, TabularData
+from power_grid_model_io.data_types.base_tabular_data import BaseTabularData, LazyDataFrame
+from power_grid_model_io.data_types.rawx_data import RawxData
+from power_grid_model_io.data_types.tabular_data import TabularData
 
-__all__ = ["ExtraInfo", "ExtraInfoLookup", "LazyDataFrame", "StructuredData", "TabularData"]
+__all__ = [
+    "BaseTabularData",
+    "ExtraInfo",
+    "ExtraInfoLookup",
+    "LazyDataFrame",
+    "RawxData",
+    "StructuredData",
+    "TabularData"
+]
